Move toolbar logic from ChatActivity to BaseActivity
